pyproject.toml
[tool.poetry]
name = "doctors-appointment"
version = "0.1.0"
description = ""
authors = ["BlaackWizard <kabynks@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.6"}
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.3"
alembic = "^1.14.0"
uvicorn = "^0.34.0"
pre-commit = "^4.0.1"
pydantic-settings = "^2.7.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.2.1"
email-validator = "^2.2.0"
aiohttp = "^3.11.11"
flower = "^2.0.1"
fastapi-mail = "^1.4.2"
itsdangerous = "^2.2.0"
qrcode = {extras = ["pil"], version = "^8.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
